solved
Approach
Merge and Sort: The naive approach would involve merging both arrays into a single sorted array and then finding the median. However, this approach has a time complexity of O(m+n), which does not meet the required O(log(m+n)) constraint.

Binary Search: To achieve O(log(m+n)), we can use a binary search approach. The idea is to partition the two arrays such that the left half of the merged array contains elements less than or equal to the elements in the right half. The median will then be either the maximum of the left half or the average of the maximum of the left half and the minimum of the right half, depending on whether the total number of elements is odd or even.